refactored autotests into basictests folder
