+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/geRM/autotests/basictests/DriveEncoderTest.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmodes.geRM.autotests;
+package org.firstinspires.ftc.teamcode.opmodes.geRM.autotests.basictests;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.hardware.DcMotor;
