+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/geRM/autotests/basictests/TurnTest.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmodes.geRM.autotests;
+package org.firstinspires.ftc.teamcode.opmodes.geRM.autotests.basictests;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.hardware.DcMotor;
@@ -14,8 +14,8 @@ import static org.firstinspires.ftc.teamcode.util.enums.Direction.RIGHT;
  * Created by General on 1/5/2018.
  */
 
-@Autonomous(name = "Turn Time Test")
-public class TurnTimeTest extends GeRMLinear{
+@Autonomous(name = "Turn Test")
+public class TurnTest extends GeRMLinear{
     @Override
     public void runOpMode() throws InterruptedException {
         super.initialize(Color.RED, DcMotor.RunMode.RUN_WITHOUT_ENCODER, FORWARD);
